test: add full Node E2E coverage and CI matrix on Node 22/24
Extract shared E2E scenario helpers so dist/cli.mjs runs the same subprocess tests as the Bun-source suite. Add a dedicated e2e-node CI job on Node 22 and 24.
