+++ .github/workflows/ci.yml
@@ -21,7 +21,6 @@ jobs:
           - check
           - knip
           - test
-          - test:node
           - security:scan
           - typecheck
     name: ${{ matrix.task }}
@@ -46,3 +45,32 @@ jobs:
 
       - name: Run task
         run: bun run ${{ matrix.task }}
+
+  e2e-node:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        node-version: [22, 24]
+    name: e2e-node (${{ matrix.node-version }})
+    steps:
+      - name: Checkout
+        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
+        with:
+          persist-credentials: false
+
+      - name: Setup Bun
+        uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2.2.0
+        with:
+          bun-version: 1.3.14
+
+      - name: Setup Node
+        uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
+        with:
+          node-version: ${{ matrix.node-version }}
+
+      - name: Install dependencies
+        run: bun install --frozen-lockfile
+
+      - name: Run Node E2E tests
+        run: bun run test:e2e:node
