+++ package.json
@@ -97,7 +97,8 @@
     "sync-skill": "bun run ./scripts/sync-skill.ts",
     "test": "bun test",
     "test:coverage": "bun test --coverage",
-    "test:node": "bun run build && bun test __tests__/cli.node.e2e.test.ts",
+    "test:e2e:node": "bun run build && bun test __tests__/**/*.node.e2e.test.ts",
+    "test:node": "bun run test:e2e:node",
     "typecheck": "tsgo --noEmit",
     "version-packages": "changeset version",
     "watch": "bun --watch src/cli.ts"
